reorganize open-authoring into potential packages
