+++ open-authoring/toolbar-plugins/pull-request/PRModal.tsx
@@ -3,7 +3,7 @@ import { Input, TextArea } from '@tinacms/fields'
 import { ModalBody, ModalActions, FieldMeta, useCMS } from 'tinacms'
 import styled from 'styled-components'
 import React, { useEffect, useState } from 'react'
-import { getHeadBranch } from '../../utils/repository'
+import { getHeadBranch } from '../../open-authoring/repository'
 
 const BASE_BRANCH = process.env.BASE_BRANCH
