+++ pages/_app.tsx
@@ -6,7 +6,7 @@ import { DefaultSeo } from 'next-seo'
 import data from '../content/siteConfig.json'
 import TagManager from 'react-gtm-module'
 import { GlobalStyle } from '../components/styles/GlobalStyle'
-import { OpenAuthoringProvider } from '../open-authoring/OpenAuthoringProvider'
+import { OpenAuthoringProvider } from '../open-authoring/open-authoring/OpenAuthoringProvider'
 import { Toolbar } from '../components/cms/Toolbar'
 import { BrowserStorageApi } from '../utils/plugins/BrowserStorageApi'
 import { Alerts } from '../components/layout/Alerts'
