+++ utils/github/getContent.ts
@@ -21,15 +21,3 @@ export const getContent = async (
       return err
     })
 }
-
-export const createAccessToken = (clientId, clientSecret, code, state) => {
-  return axios.post(
-    `https://github.com/login/oauth/access_token`,
-    qs.stringify({
-      client_id: clientId,
-      client_secret: clientSecret,
-      code: code,
-      state,
-    })
-  )
-}
