+++ utils/plugins/github-api/GithubError.ts
